+++ lib/persistance/connection/dao/_base.dart
@@ -2,7 +2,7 @@ import '../../model/_entity.dart';
 
 abstract class BaseDao<T extends Entity> {
 
-	T? get(int? id);
+	T? get(int id);
 	List<T> getAll();
 	bool insert(T t);
 	bool update(T t);