+++ lib/persistance/model/auth.dart
@@ -8,16 +8,19 @@ class Auth extends Entity {
 
 	Auth(this.id, this.username, this.password, this.salt, this.admin, this.user);
 
-	factory Auth.fromServer(final Map<String, dynamic> map) => Auth(
-		map['ID'], map['Username'], map['Password'], map['Salt'],
-		map['Admin'] != 0 ? true : false, DaoFactory.instance.userDao.get(map['UserID'])
-	);
+	factory Auth.fromServer(final Map<String, dynamic> map) {
+		int? userID = map['UserID'];
+		return Auth(
+			map['ID'], map['Username'], map['Password'], map['Salt'],
+			map['Admin'] != 0 ? true : false, userID != null ? DaoFactory.instance.userDao.get(map['UserID'])! : null
+		);
+	}
 
 	factory Auth.exp(int id, String username, String password, bool admin, int? userID) {
 		Random random = Random.secure();
 		int salt = random.nextInt(0x7FFFFFFF);
 		int password_hash = password.hashCode ^ salt;
-		return Auth(id, username, password_hash, salt, admin, DaoFactory.instance.userDao.get(userID));
+		return Auth(id, username, password_hash, salt, admin, userID != null ? DaoFactory.instance.userDao.get(userID) : null);
 	}
 
 	factory Auth.fromJson(final Map<String, dynamic> map, final Random random) {
@@ -26,7 +29,7 @@ class Auth extends Entity {
 		int? userID = map['UserID'];
 		return Auth(
 			map['ID'] ?? -1, map['Username'], password, salt,
-			map['Admin'], userID != null ? DaoFactory.instance.userDao.get(map['UserID'])! : null
+			map['Admin'] ?? false, userID != null ? DaoFactory.instance.userDao.get(map['UserID'])! : null
 		);
 	}