+++ lib/server/router.dart
@@ -1,3 +1,4 @@
+import 'dart:collection';
 import 'dart:io';
 
 import 'package:happy_hour_server/server/filter_chain.dart';
@@ -28,6 +29,24 @@ class Router {
 		for (FilterChain filterChain in _routes) {
 			if (request.uri.path.startsWith(filterChain.path)) {
 				return filterChain.handle(request);
+			}
+		}
+		return Response.notImplemented(request);
+	}
+
+}
+
+class RouterV2 {
+
+	// LinkedHashMap ITERATES IN INSERTION ORDER, CRUCIAL
+	final Map<String, RequestDispatcher> _routes = LinkedHashMap<String, RequestDispatcher>();
+
+	void addRoute(final String path, final RequestDispatcher dispatcher) => _routes[path] = dispatcher;
+
+	Future handle(HttpRequest request) {
+		for (final MapEntry<String, RequestDispatcher> route in _routes.entries) {
+			if (request.uri.path.startsWith(route.key)) {
+				return route.value.handle(request);
 			}
 		}
 		return Response.notImplemented(request);