+++ lib/server/runner.dart
@@ -11,3 +11,12 @@ Future runServer(Server srv) => HttpServer.bind(InternetAddress.anyIPv4, 8080)
 		}
 	)
 	.catchError((e) => Logger.e(e));
+
+Future runServerV2(ServerV2 srv) => HttpServer.bind(InternetAddress.anyIPv4, 8080)
+	.then(
+		(HttpServer server) {
+			srv.onInit();
+			server.listen(srv.router.handle);
+		}
+	)
+	.catchError((e) => Logger.e(e));