+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 21.3.18
+
+- Dart 2.12.1 required
+- More lightweight server dispatcher
+- Get image from the server
+- [TODO] Request filters by method (not crucial, only removes code from actual methods)
+
 ## 21.3.17-1
 
 - Added V2 API: more lightweight and less chaotic